Fix remounting of dialog when opened at the first time Fixes #112
