+++ packages/ui/package.json
@@ -24,6 +24,7 @@
   },
   "dependencies": {
     "@gc/i18n": "workspace:*",
+    "@gc/react-kit": "workspace:*",
     "@gc/utils": "workspace:*",
     "@radix-ui/react-checkbox": "^1.3.3",
     "@radix-ui/react-dialog": "^1.1.14",
