+++ packages/ui/src/components/ui/dialog/dialog-button.tsx
@@ -1,0 +1,15 @@
+import { Button, ButtonProps } from "../button";
+import { useDialog } from "./use-dialog";
+
+export function DialogButton({
+  dialogContent,
+  ...props
+}: { dialogContent: React.ReactNode } & ButtonProps) {
+  const { openDialog } = useDialog();
+  return (
+    <Button
+      onClick={() => openDialog(dialogContent)}
+      {...props}
+    />
+  );
+}
