+++ packages/ui/src/components/ui/dialog/index.ts
@@ -1,3 +1,4 @@
 export * from "./dialog-provider";
 export * from "./use-dialog";
 export * from "./dialog-layout";
+export * from "./dialog-button";
